third_party/WebKit/LayoutTests/http/tests/loading/bad-scheme-subframe.html

<script>
if (window.testRunner)
    testRunner.dumpAsText();

window.onload = function() {
  var i = document.createElement("iframe");
  i.name = "f1";
  i.src = "this-is-definitely-a-bad-uri-scheme:";
  document.body.appendChild(i);
}
</script>
This is a test of load callbacks. It is only useful inside the regression test tool.<br>
